chore(federation): move RebasedFragments into the query_plan module
